+++ behaviour-detector/src/test/java/au/gov/amsa/geo/adhoc/CountCrossingsIntoRegionMain.java
@@ -12,8 +12,6 @@ import java.util.regex.Pattern;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import rx.Observable;
-import rx.functions.Func2;
 import au.gov.amsa.ais.ShipTypeDecoder;
 import au.gov.amsa.gt.Shapefile;
 import au.gov.amsa.navigation.ShipStaticData;
@@ -21,6 +19,8 @@ import au.gov.amsa.navigation.ShipStaticData.Info;
 import au.gov.amsa.risky.format.BinaryFixes;
 import au.gov.amsa.risky.format.Fix;
 import au.gov.amsa.util.Files;
+import rx.Observable;
+import rx.functions.Func2;
 
 public class CountCrossingsIntoRegionMain {
 
@@ -34,7 +34,7 @@ public class CountCrossingsIntoRegionMain {
         region.contains(0, 0);
 
         Func2<Fix, Fix, Integer> ascendingTime = (a, b) -> ((Long) a.time()).compareTo(b.time());
-        Map<Long, Info> ships = ShipStaticData.getMapFromResource("/ship-data.txt");
+        Map<Integer, Info> ships = ShipStaticData.getMapFromResource("/ship-data.txt");
         Pattern pattern = Pattern.compile(".*\\.track");
         List<File> files = Files.find(new File("/media/an/binary-fixes-5-minute/2012"), pattern);
         files.addAll(Files.find(new File("/media/an/binary-fixes-5-minute/2013"), pattern));
@@ -42,40 +42,36 @@ public class CountCrossingsIntoRegionMain {
         files.addAll(Files.find(new File("/media/an/binary-fixes-5-minute/2015"), pattern));
         log.info("files=" + files.size());
 
-        int count = Observable
-                .from(files)
+        int count = Observable.from(files)
                 // .doOnNext(System.out::println)
                 .concatMap(file -> detectCrossings(file, region))
                 // count out to in crossings
-                .toSortedList(ascendingTime)
-                .flatMap(o -> Observable.from(o))
+                .toSortedList(ascendingTime).flatMap(o -> Observable.from(o))
                 // log
-                .doOnNext(
-                        fix -> {
-                            Info info = ships.get(fix.mmsi());
-                            if (info != null) {
-                                String type;
-                                if (info.shipType.isPresent())
-                                    type = ShipTypeDecoder.getShipType(info.shipType.get());
-                                else
-                                    type = "UNKNOWN";
-                                String t = DateTimeFormatter.ISO_DATE_TIME.format(ZonedDateTime
-                                        .ofInstant(Instant.ofEpochMilli(fix.time()),
-                                                ZoneId.of("UTC")));
-                                System.out.println(t + ", mmsi=" + fix.mmsi() + ", class="
-                                        + fix.aisClass().name() + ", lengthMetres="
-                                        + info.lengthMetres().orNull() + ", type=" + type
-                                        + ", lat=" + fix.lat() + ", lon=" + fix.lon());
-                            }
-                        }).count().toBlocking().single();
+                .doOnNext(fix -> {
+                    Info info = ships.get(fix.mmsi());
+                    if (info != null) {
+                        String type;
+                        if (info.shipType.isPresent())
+                            type = ShipTypeDecoder.getShipType(info.shipType.get());
+                        else
+                            type = "UNKNOWN";
+                        String t = DateTimeFormatter.ISO_DATE_TIME.format(ZonedDateTime
+                                .ofInstant(Instant.ofEpochMilli(fix.time()), ZoneId.of("UTC")));
+                        System.out.println(t + ", mmsi=" + fix.mmsi() + ", class="
+                                + fix.aisClass().name() + ", lengthMetres="
+                                + info.lengthMetres().orNull() + ", type=" + type + ", lat="
+                                + fix.lat() + ", lon=" + fix.lon());
+                    }
+                }).count().toBlocking().single();
         System.out.println("count=" + count);
     }
 
     private static Observable<Fix> detectCrossings(File file, Shapefile region) {
         return BinaryFixes.from(file)
-        // log
-        // .lift(Logging.<Fix>
-        // logger().showCount().every(100000).log())
+                // log
+                // .lift(Logging.<Fix>
+                // logger().showCount().every(100000).log())
                 .filter(fix -> fix.mmsi() != 0)
                 // add in or out of region
                 .map(fix -> new FixAndRegion(fix, region.contains(fix.lat(), fix.lon())))
